+++ lib/hexapdf/font_loader.rb
@@ -79,6 +79,7 @@ module HexaPDF
 
     autoload(:Standard14, 'hexapdf/font_loader/standard14')
     autoload(:FromConfiguration, 'hexapdf/font_loader/from_configuration')
+    autoload(:FromFile, 'hexapdf/font_loader/from_file')
 
   end